% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_book}
\alias{setup_book}
\title{Setup book}
\usage{
setup_book()
}
\description{
Creates book file structure. It will contain all the files that you need
to compile your new book. You should start by running \code{Ctrl+Shift+B}
(build) on index.Rmd using RStudio. If it runs successfully, your output will
be in the docs folder
}
\examples{
\dontrun{
bookTemplate::setup_book()
}
}
